refactor: validator 제거, 도메인 종속적 validate로 변경
