+++ src/main/java/racingcar/domain/RaceRound.java
@@ -1,17 +1,27 @@
 package racingcar.domain;
 
-import racingcar.validator.RaceRoundInputValidator;
-import racingcar.validator.InputValidator;
-
 public class RaceRound {
 
+    private static final String Race_Round_IS_DIGIT = "입력값이 숫자가 아닙니다.";
+
     private int rounds;
-    private InputValidator validator = new RaceRoundInputValidator();
 
     public RaceRound(String stringRound) {
-        validator.validateInput(stringRound);
+        validateInput(stringRound);
         int round = Integer.parseInt(stringRound);
         rounds = round;
+    }
+
+    public void validateInput(String input) {
+        keepDigit(input);
+    }
+
+    public void keepDigit(String input) {
+        try {
+            Integer.parseInt(input);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException(Race_Round_IS_DIGIT);
+        }
     }
 
     public int getRounds() {
