--- src/main/java/racingcar/validator/InputValidator.java
@@ -1,5 +1,0 @@
-package racingcar.validator;
-
-public interface InputValidator {
-    public void validateInput(String input);
-}
